voyage-system/voyage-app-backend/requirements.txt
fastapi
uvicorn
python-dotenv
cryptography
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
